Refactoring table column code
- Moves all column definitions to columns.js
- Moves helpers from tables.js and columns.js to a new column-helpers.js
- Creates new renderCandidateColumn and renderCommitteeColumn to
consolidate render callback code
