+++ static/js/modules/columns.js
@@ -2,30 +2,15 @@
 
 var _ = require('underscore');
 
+var columnHelpers = require('./column-helpers');
 var tables = require('./tables');
 var helpers = require('./helpers');
 var decoders = require('./decoders');
 
-var sizeInfo = {
-  0: {limits: [0, 199.99], label: 'Under $200'},
-  200: {limits: [200, 499.99], label: '$200—$499'},
-  500: {limits: [500, 999.99], label: '$500—$999'},
-  1000: {limits: [1000, 1999.99], label: '$1000—$1999'},
-  2000: {limits: [2000, null], label: 'Over $2000'},
-};
 
-function getSizeParams(size) {
-  var limits = sizeInfo[size].limits;
-  var params = {is_individual: 'true'};
-  if (limits[0] !== null) {
-    params.min_amount = helpers.currency(limits[0]);
-  }
-  if (limits[1] !== null) {
-    params.max_amount = helpers.currency(limits[1]);
-  }
-  return params;
-}
-
+var dateColumn = columnHelpers.formattedColumn(helpers.datetime, {orderSequence: ['desc', 'asc']});
+var currencyColumn = columnHelpers.formattedColumn(helpers.currency, {orderSequence: ['desc', 'asc']});
+var barCurrencyColumn = columnHelpers.barColumn(helpers.currency);
 
 var supportOpposeColumn = {
   data: 'support_oppose_indicator',
@@ -42,8 +27,150 @@ var amendmentIndicatorColumn = {
   },
 };
 
+var modalTriggerColumn = {
+  className: 'all u-no-padding',
+  width: '20px',
+  orderable: false,
+  render: function(data, type, row, meta) {
+    return tables.MODAL_TRIGGER_HTML;
+  }
+};
+
+var renderCandidateColumn = function(data, type, row, meta) {
+  if (data) {
+    var query = row.election_years ? tables.getCycle(row.election_years, meta) : null;
+    return columnHelpers.buildEntityLink(data, helpers.buildAppUrl(['candidate', row.candidate_id], query), 'candidate');
+  } else {
+    return '';
+  }
+};
+
+var renderCommitteeColumn = function(data, type, row, meta) {
+  if (data) {
+    var query = row.cycles ? tables.getCycle(row.cycles, meta) : null;
+    return columnHelpers.buildEntityLink(
+      data,
+      helpers.buildAppUrl(['committee', row.committee_id], query),
+      'committee');
+  } else {
+    return '';
+  }
+};
+
+var candidates = [
+  {data: 'name', className: 'all', width: '280px', render: renderCandidateColumn},
+  {data: 'office_full', className: 'min-tablet hide-panel'},
+  {
+    data: 'election_years',
+    className: 'min-tablet',
+    render: function(data, type, row, meta) {
+      return tables.yearRange(_.first(data), _.last(data));
+    }
+  },
+  {data: 'party_full', className: 'min-tablet hide-panel'},
+  {data: 'state', className: 'min-desktop hide-panel'},
+  {data: 'district', className: 'min-desktop hide-panel'},
+  modalTriggerColumn
+];
+
+var candidateOffice = {
+  candidate:   {data: 'name', className: 'all', width: '280px', render: renderCandidateColumn},
+  party: {data: 'party_full', className: 'min-tablet hide-panel'},
+  state: {data: 'state', className: 'min-desktop hide-panel'},
+  district: {data: 'district', className: 'min-desktop hide-panel'},
+  receipts: currencyColumn({data: 'receipts', className: 'min-tablet'}),
+  disbursements: currencyColumn({data: 'disbursements', className: 'min-tablet'}),
+  trigger: modalTriggerColumn
+};
+
+var committees = [
+  {data: 'name', className: 'all', width: '280px', render: renderCommitteeColumn},
+  {data: 'treasurer_name', className: 'min-desktop hide-panel'},
+  {data: 'state', className: 'min-desktop hide-panel', width: '60px'},
+  {data: 'party_full', className: 'min-desktop hide-panel'},
+  dateColumn({data: 'first_file_date', className: 'min-tablet hide-panel'}),
+  {data: 'committee_type_full', className: 'min-tablet hide-panel'},
+  {data: 'designation_full', className: 'min-tablet hide-panel'},
+  {data: 'organization_type_full', className: 'min-desktop hide-panel'},
+  modalTriggerColumn
+];
+
+var communicationCosts = [
+  {
+    data: 'committee_name',
+    orderable: false,
+    className: 'min-desktop',
+    render: renderCommitteeColumn,
+  },
+  currencyColumn({data: 'transaction_amount', className: 'min-tablet'}),
+  supportOpposeColumn,
+  {
+    data: 'candidate_name',
+    orderable: false,
+    className: 'min-desktop hide-panel',
+    render: renderCandidateColumn
+  },
+  dateColumn({data: 'transaction_date', className: 'min-tablet hide-panel-tablet'}),
+  modalTriggerColumn
+];
+
+var disbursements = [
+  {
+    data: 'recipient_name',
+    orderable: false,
+    className: 'all',
+    width: '200px',
+    render: function(data, type, row, meta) {
+      var committee = row.recipient_committee;
+      if (committee) {
+        return columnHelpers.buildEntityLink(
+          committee.name,
+          helpers.buildAppUrl(['committee', committee.committee_id]),
+          'committee'
+        );
+      } else {
+        return data;
+      }
+    }
+  },
+  {data: 'recipient_state', width: '80px', orderable: false, className: 'min-desktop hide-panel'},
+  currencyColumn({data: 'disbursement_amount', className: 'min-tablet hide-panel-tablet'}),
+  dateColumn({data: 'disbursement_date', className: 'min-tablet hide-panel-tablet'}),
+  {data: 'disbursement_description', className: 'min-desktop hide-panel', orderable: false},
+  {
+    data: 'committee',
+    orderable: false,
+    className: 'min-tablet hide-panel',
+    render: renderCommitteeColumn
+  },
+  modalTriggerColumn
+];
+
+var electioneeringCommunications = [
+  {
+    data: 'committee_name',
+    orderable: false,
+    className: 'min-desktop',
+    render: renderCommitteeColumn
+  },
+  currencyColumn({data: 'disbursement_amount', className: 'min-tablet'}),
+  {
+    data: 'number_of_candidates',
+    className: 'min-tablet',
+  },
+  currencyColumn({data: 'calculated_candidate_share', className: 'min-tablet'}),
+  {
+    data: 'candidate_name',
+    orderable: false,
+    className: 'min-desktop hide-panel hide-panel-tablet',
+    render: renderCandidateColumn
+  },
+  dateColumn({data: 'disbursement_date', className: 'min-tablet hide-panel'}),
+  modalTriggerColumn
+];
+
 var filings = {
-  pdf_url: tables.urlColumn('pdf_url', {data: 'document_description', className: 'all', orderable: false}),
+  pdf_url: columnHelpers.urlColumn('pdf_url', {data: 'document_description', className: 'all', orderable: false}),
   filer_name: {
     data: 'committee_id',
     className: 'all',
@@ -51,13 +178,13 @@ var filings = {
     render: function(data, type, row, meta) {
       var cycle = tables.getCycle([row.cycle], meta);
       if (row.candidate_name) {
-        return tables.buildEntityLink(
+        return columnHelpers.buildEntityLink(
           row.candidate_name,
           helpers.buildAppUrl(['candidate', row.candidate_id], cycle),
           'candidate'
         );
       } else if (row.committee_name) {
-        return tables.buildEntityLink(
+        return columnHelpers.buildEntityLink(
           row.committee_name,
           helpers.buildAppUrl(['committee', row.committee_id], cycle),
           'committee'
@@ -68,34 +195,99 @@ var filings = {
     },
   },
   amendment_indicator: amendmentIndicatorColumn,
-  receipt_date: tables.dateColumn({data: 'receipt_date', className: 'min-tablet'}),
-  coverage_end_date: tables.dateColumn({data: 'coverage_end_date', className: 'min-tablet hide-panel', orderable: false}),
-  total_receipts: tables.currencyColumn({data: 'total_receipts', className: 'min-tablet hide-panel'}),
-  total_disbursements: tables.currencyColumn({data: 'total_disbursements', className: 'min-tablet hide-panel'}),
-  total_independent_expenditures: tables.currencyColumn({data: 'total_independent_expenditures', className: 'min-tablet hide-panel'}),
-  modal_trigger: {
-    className: 'all u-no-padding',
-    width: '20px',
-    orderable: false,
-    render: function(data, type, row, meta) {
-      return row.form_type && ['F3', 'F3P', 'F3X'].indexOf(row.form_type) !== -1 ?
-        tables.MODAL_TRIGGER_HTML :
-        '';
-    }
-  }
+  receipt_date: dateColumn({data: 'receipt_date', className: 'min-tablet'}),
+  coverage_end_date: dateColumn({data: 'coverage_end_date', className: 'min-tablet hide-panel', orderable: false}),
+  total_receipts: currencyColumn({data: 'total_receipts', className: 'min-tablet hide-panel'}),
+  total_disbursements: currencyColumn({data: 'total_disbursements', className: 'min-tablet hide-panel'}),
+  total_independent_expenditures: currencyColumn({data: 'total_independent_expenditures', className: 'min-tablet hide-panel'}),
+  modal_trigger: modalTriggerColumn
 };
 
-function getColumns(columns, keys) {
-  return _.map(keys, function(key) {
-    return columns[key];
-  });
-}
+var independentExpenditures = [
+  {
+    data: 'committee',
+    orderable: false,
+    className: 'all',
+    render: renderCommitteeColumn
+  },
+  currencyColumn({data: 'expenditure_amount', className: 'min-tablet'}),
+  {
+    data: 'candidate_name',
+    orderable: false,
+    className: 'min-tablet hide-panel',
+    render: function(data, type, row, meta) {
+      if (row.candidate_id) {
+        return columnHelpers.buildEntityLink(
+          data,
+          helpers.buildAppUrl(['candidate', row.candidate_id], tables.getCycle(row, meta)),
+          'candidate'
+        );
+      } else {
+        return row.candidate_name;
+      }
+    }
+  },
+  _.extend({}, supportOpposeColumn, {className: 'min-tablet'}),
+  dateColumn({data: 'expenditure_date', className: 'min-desktop hide-panel-tablet'}),
+  columnHelpers.urlColumn('pdf_url', {data: 'expenditure_description', className: 'min-desktop hide-panel', orderable: false}),
+  modalTriggerColumn
+];
+
+var receipts = [
+  {
+    data: 'contributor',
+    orderable: false,
+    className: 'all',
+    width: '200px',
+    render: function(data, type, row, meta) {
+      if (data && row.receipt_type !== helpers.globals.EARMARKED_CODE) {
+        return columnHelpers.buildEntityLink(
+          data.name,
+          helpers.buildAppUrl(['committee', data.committee_id]),
+          'committee'
+        );
+      } else {
+        return row.contributor_name;
+      }
+    }
+  },
+  {data: 'contributor_state', orderable: false, className: 'min-desktop hide-panel'},
+  {data: 'contributor_employer', orderable: false, className: 'min-desktop hide-panel'},
+  currencyColumn({data: 'contribution_receipt_amount', className: 'min-tablet'}),
+  dateColumn({data: 'contribution_receipt_date', className: 'min-tablet hide-panel-tablet'}),
+  {
+    data: 'committee',
+    orderable: false,
+    className: 'min-desktop hide-panel',
+    render: function(data, type, row, meta) {
+      if (data) {
+        return columnHelpers.buildEntityLink(
+          data.name,
+          helpers.buildAppUrl(['committee', data.committee_id]),
+          'committee'
+        );
+      } else {
+        return '';
+      }
+    }
+  },
+  modalTriggerColumn
+];
+
 
 module.exports = {
-  filings: filings,
-  sizeInfo: sizeInfo,
-  getColumns: getColumns,
-  getSizeParams: getSizeParams,
+  dateColumn: dateColumn,
+  currencyColumn: currencyColumn,
+  barCurrencyColumn: barCurrencyColumn,
   supportOpposeColumn: supportOpposeColumn,
-  amendmentIndicatorColumn: amendmentIndicatorColumn
+  amendmentIndicatorColumn: amendmentIndicatorColumn,
+  candidates: candidates,
+  candidateOffice: candidateOffice,
+  committees: committees,
+  communicationCosts: communicationCosts,
+  disbursements: disbursements,
+  electioneeringCommunications: electioneeringCommunications,
+  independentExpenditures: independentExpenditures,
+  filings: filings,
+  receipts: receipts
 };
