+++ static/js/modules/tables.js
@@ -76,118 +76,6 @@ function getCycle(value, meta) {
   }
 }
 
-function buildEntityLink(data, url, category, opts) {
-  opts = opts || {};
-  var anchor = document.createElement('a');
-  anchor.textContent = data;
-  anchor.setAttribute('href', url);
-  anchor.setAttribute('title', data);
-  anchor.setAttribute('data-category', category);
-  anchor.classList.add('single-link');
-
-  if (opts.isIncumbent) {
-    anchor.classList.add('is-incumbent');
-  }
-
-  return anchor.outerHTML;
-}
-
-function buildAggregateUrl(cycle) {
-  var dates = helpers.cycleDates(cycle);
-  return {
-    min_date: dates.min,
-    max_date: dates.max
-  };
-}
-
-function buildTotalLink(path, getParams) {
-  return function(data, type, row, meta) {
-    data = data || 0;
-    var params = getParams(data, type, row, meta);
-    var span = document.createElement('div');
-    span.setAttribute('data-value', data);
-    span.setAttribute('data-row', meta.row);
-    if (params) {
-      var link = document.createElement('a');
-      link.textContent = helpers.currency(data);
-      link.setAttribute('title', 'View individual transactions');
-      var uri = helpers.buildAppUrl(path, _.extend(
-        {committee_id: row.committee_id},
-        buildAggregateUrl(_.extend({}, row, params).cycle),
-        params
-      ));
-      link.setAttribute('href', uri);
-      span.appendChild(link);
-    } else {
-      span.textContent = helpers.currency(data);
-    }
-    return span.outerHTML;
-  };
-}
-
-function formattedColumn(formatter, defaultOpts) {
-  defaultOpts = defaultOpts || {};
-  return function(opts) {
-    return _.extend({}, defaultOpts, {
-      render: function(data, type, row, meta) {
-        return formatter(data, type, row, meta);
-      }
-    }, opts);
-  };
-}
-
-function barColumn(formatter) {
-  formatter = formatter || function(value) { return value; };
-  return function(opts) {
-    return _.extend({
-      orderSequence: ['desc', 'asc'],
-      render: function(data, type, row, meta) {
-        var span = document.createElement('div');
-        span.textContent = formatter(_.max([data, 0]));
-        span.setAttribute('data-value', data || 0);
-        span.setAttribute('data-row', meta.row);
-        return span.outerHTML;
-      }
-    }, opts);
-  };
-}
-
-function urlColumn(attr, opts) {
-  return _.extend({
-    render: function(data, type, row, meta) {
-      if (row[attr]) {
-        var anchor = document.createElement('a');
-        anchor.textContent = data;
-        anchor.setAttribute('href', row[attr]);
-        anchor.setAttribute('target', '_blank');
-        return anchor.outerHTML;
-      } else {
-        return data;
-      }
-    }
-  }, opts);
-}
-
-var dateColumn = formattedColumn(helpers.datetime, {orderSequence: ['desc', 'asc']});
-var currencyColumn = formattedColumn(helpers.currency, {orderSequence: ['desc', 'asc']});
-var barCurrencyColumn = barColumn(helpers.currency);
-
-var candidateColumn = formattedColumn(function(data, type, row) {
-  if (row) {
-    return buildEntityLink(row.candidate_name, helpers.buildAppUrl(['candidate', row.candidate_id]), 'candidate');
-  } else {
-    return '';
-  }
-});
-
-var committeeColumn = formattedColumn(function(data, type, row) {
-  if (row) {
-    return buildEntityLink(row.committee_name, helpers.buildAppUrl(['committee', row.committee_id]), 'committee');
-  } else {
-    return '';
-  }
-});
-
 function mapSort(order, columns) {
   return _.map(order, function(item) {
     var name = columns[item.column].data;
@@ -613,14 +501,6 @@ module.exports = {
   browseDOM: browseDOM,
   yearRange: yearRange,
   getCycle: getCycle,
-  buildTotalLink: buildTotalLink,
-  buildEntityLink: buildEntityLink,
-  candidateColumn: candidateColumn,
-  committeeColumn: committeeColumn,
-  currencyColumn: currencyColumn,
-  urlColumn: urlColumn,
-  barCurrencyColumn: barCurrencyColumn,
-  dateColumn: dateColumn,
   barsAfterRender: barsAfterRender,
   modalRenderRow: modalRenderRow,
   modalRenderFactory: modalRenderFactory,
