+++ static/js/pages/electioneering-communications.js
@@ -4,58 +4,13 @@ var $ = require('jquery');
 
 var tables = require('../modules/tables');
 var helpers = require('../modules/helpers');
+var columnHelpers = require('../modules/column-helpers');
 var columns = require('../modules/columns');
 
 var FilterPanel = require('fec-style/js/filter-panel').FilterPanel;
 var filterTags = require('fec-style/js/filter-tags');
 
 var electioneeringTemplate = require('../../templates/electioneering-communications.hbs');
-
-var columns = [
-  {
-    data: 'committee_name',
-    orderable: false,
-    className: 'min-desktop',
-    render: function(data, type, row, meta) {
-      if (data) {
-        return tables.buildEntityLink(
-          data,
-          helpers.buildAppUrl(['committee', row.committee_id]),
-          'committee'
-        );
-      } else {
-        return '';
-      }
-    }
-  },
-  tables.currencyColumn({data: 'disbursement_amount', className: 'min-tablet'}),
-  {
-    data: 'number_of_candidates',
-    className: 'min-tablet',
-  },
-  tables.currencyColumn({data: 'calculated_candidate_share', className: 'min-tablet'}),
-  {
-    data: 'candidate_name',
-    orderable: false,
-    className: 'min-desktop hide-panel hide-panel-tablet',
-    render: function(data, type, row, meta) {
-      return tables.buildEntityLink(
-        data,
-        helpers.buildAppUrl(['candidate', row.candidate_id]),
-        'candidate'
-      );
-    }
-  },
-  tables.dateColumn({data: 'disbursement_date', className: 'min-tablet hide-panel'}),
-  {
-    className: 'all u-no-padding',
-    width: '20px',
-    orderable: false,
-    render: function(data, type, row, meta) {
-      return tables.MODAL_TRIGGER_HTML;
-    }
-  }
-];
 
 $(document).ready(function() {
   var $table = $('#results');
@@ -66,7 +21,7 @@ $(document).ready(function() {
     title: 'Electioneering communications',
     path: ['electioneering'],
     panel: filterPanel,
-    columns: columns,
+    columns: columns.electioneeringCommunications,
     rowCallback: tables.modalRenderRow,
     useExport: true,
     order: [[3, 'desc']],
